Add error handeling pages (primarily for 404)
Now you can add special error pages for http error codes.
Currently it works only for 404 errors. If a route is called, which
isn't defined, the framwork gave our a generic 404 error page.
Now you can redirect this error to a controller, which allowes you to
create a special error handling for that case yourself.
you just need to add a route "404 => MyErrorController" to your routing
list. If it's not defined, noting will change and the old error message
will be returned as beforehand.