Correct HTTP status code on redirect
With the release of RFC 7231 in June 2014 the HTTP status code 302 - Moved
Temporarily does no longer exist in this form and has been replaced by
302 - Found, which causes the browser to use the same request method to
call the new resource.
Which can cause unwanted deletions of resouces, when a child resource is
deleted and the browser gets redirected to the parent resource.
Instead we use now the status code 303 - See Others, which specifies, that
the referred resource has to be called using the GET method.